forge fmt

+++ src/SimpleVotingSystem.sol
@@ -9,22 +9,27 @@ contract SimpleVotingSystem is AccessControl {
     bytes32 public constant FOUNDER_ROLE = keccak256("FOUNDER_ROLE");
 
     struct Candidate {
-        uint id;
+        uint256 id;
         string name;
-        uint voteCount;
+        uint256 voteCount;
     }
 
-    mapping(uint => Candidate) public candidates;
+    mapping(uint256 => Candidate) public candidates;
     mapping(address => bool) public voters;
-    uint[] private candidateIds;
+    uint256[] private candidateIds;
 
-    enum WorkflowStatus { REGISTER_CANDIDATES, FOUND_CANDIDATES, VOTE, COMPLETED }
+    enum WorkflowStatus {
+        REGISTER_CANDIDATES,
+        FOUND_CANDIDATES,
+        VOTE,
+        COMPLETED
+    }
     WorkflowStatus public workflowStatus;
 
-    mapping(uint => uint256) public candidateFunds;
+    mapping(uint256 => uint256) public candidateFunds;
 
     uint256 public voteStartSetAt;
-    
+
     SimpleVotingNFT public votingNFT;
 
     constructor(address _votingNFT) {
@@ -35,12 +40,12 @@ contract SimpleVotingSystem is AccessControl {
     function addCandidate(string memory _name) public onlyRole(ADMIN_ROLE) {
         require(workflowStatus == WorkflowStatus.REGISTER_CANDIDATES, "Wrong phase");
         require(bytes(_name).length > 0, "Candidate name cannot be empty");
-        uint candidateId = candidateIds.length + 1;
+        uint256 candidateId = candidateIds.length + 1;
         candidates[candidateId] = Candidate(candidateId, _name, 0);
         candidateIds.push(candidateId);
     }
 
-    function vote(uint _candidateId) public {
+    function vote(uint256 _candidateId) public {
         require(workflowStatus == WorkflowStatus.VOTE, "Wrong phase");
         require(votingNFT.balanceOf(msg.sender) == 0, "You have already voted (NFT detected)");
         require(_candidateId > 0 && _candidateId <= candidateIds.length, "Invalid candidate ID");
@@ -51,17 +56,17 @@ contract SimpleVotingSystem is AccessControl {
         votingNFT.mint(msg.sender);
     }
 
-    function getTotalVotes(uint _candidateId) public view returns (uint) {
+    function getTotalVotes(uint256 _candidateId) public view returns (uint256) {
         require(_candidateId > 0 && _candidateId <= candidateIds.length, "Invalid candidate ID");
         return candidates[_candidateId].voteCount;
     }
 
-    function getCandidatesCount() public view returns (uint) {
+    function getCandidatesCount() public view returns (uint256) {
         return candidateIds.length;
     }
 
     // Optional: Function to get candidate details by ID
-    function getCandidate(uint _candidateId) public view returns (Candidate memory) {
+    function getCandidate(uint256 _candidateId) public view returns (Candidate memory) {
         require(_candidateId > 0 && _candidateId <= candidateIds.length, "Invalid candidate ID");
         return candidates[_candidateId];
     }
@@ -77,26 +82,26 @@ contract SimpleVotingSystem is AccessControl {
         _grantRole(FOUNDER_ROLE, a);
     }
 
-    function fundCandidate(uint candidateId) external payable onlyRole(FOUNDER_ROLE) {
+    function fundCandidate(uint256 candidateId) external payable onlyRole(FOUNDER_ROLE) {
         require(workflowStatus == WorkflowStatus.FOUND_CANDIDATES, "Wrong phase");
         candidateFunds[candidateId] += msg.value;
     }
 
-    function getWinner() public view returns (uint winnerId, string memory winnerName, uint winnerVoteCount) {
+    function getWinner() public view returns (uint256 winnerId, string memory winnerName, uint256 winnerVoteCount) {
         require(workflowStatus == WorkflowStatus.COMPLETED, "Voting not completed yet");
         require(candidateIds.length > 0, "No candidates registered");
-        
-        uint maxVotes = 0;
-        uint winningCandidateId = 0;
-        
-        for (uint i = 0; i < candidateIds.length; i++) {
-            uint candidateId = candidateIds[i];
+
+        uint256 maxVotes = 0;
+        uint256 winningCandidateId = 0;
+
+        for (uint256 i = 0; i < candidateIds.length; i++) {
+            uint256 candidateId = candidateIds[i];
             if (candidates[candidateId].voteCount > maxVotes) {
                 maxVotes = candidates[candidateId].voteCount;
                 winningCandidateId = candidateId;
             }
         }
-        
+
         require(winningCandidateId > 0, "No winner found");
         Candidate memory winner = candidates[winningCandidateId];
         return (winner.id, winner.name, winner.voteCount);
